M2RS-26 refactor: feign 설정 클래스 분리
