+++ m2rs-meeting-room-service/src/main/java/com/m2rs/meetingroomservice/configure/AppConfiguration.java
@@ -2,18 +2,15 @@ package com.m2rs.meetingroomservice.configure;
 
 import com.m2rs.core.security.model.JwtClaimInfo;
 import com.m2rs.core.security.utils.JwtUtils;
-import feign.Logger;
 import io.jsonwebtoken.Claims;
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.http.HttpHeaders;
-import org.springframework.web.client.RestTemplate;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
@@ -47,17 +44,6 @@ public class AppConfiguration {
             return Optional.ofNullable(userId);
 
         };
-    }
-
-    @Bean
-    @LoadBalanced
-    public RestTemplate getRestTemplate() {
-        return new RestTemplate();
-    }
-
-    @Bean
-    public Logger.Level getLevel() {
-        return Logger.Level.FULL;
     }
 
 }
