+++ m2rs-meeting-room-service/src/main/java/com/m2rs/meetingroomservice/configure/fegin/FeignConfiguration.java
@@ -1,0 +1,23 @@
+package com.m2rs.meetingroomservice.configure.fegin;
+
+import feign.Logger;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class FeignConfiguration {
+
+    @Bean
+    @LoadBalanced
+    public RestTemplate getRestTemplate() {
+        return new RestTemplate();
+    }
+
+    @Bean
+    public Logger.Level getLevel() {
+        return Logger.Level.FULL;
+    }
+
+}
